requirements.txt
aiogram==2.21
python-environ==0.4.54